code/Preprocessing service/requirements.txt
kafka-python
elasticsearch
pandas
numpy
spacy
bs4
requests
fuzzywuzzy
langdetect
python-Levenshtein
pycountry
python-restcountries
lxml
flask
flask_cors
python-whois
tld
tldextract
unidecode

